# Improve performance when writes exceed max-values-per-tag or max-series