--- a/pkg/bytesutil/bytesutil_test.go
+++ b/pkg/bytesutil/bytesutil_test.go
@@ -34,6 +34,54 @@ func TestSearchBytesFixed(t *testing.T) {
 	}
 }
 
+func TestSearchBytes(t *testing.T) {
+	in := toByteSlices("bbb", "ccc", "eee", "fff", "ggg", "hhh")
+	tests := []struct {
+		name string
+		x    string
+		exp  int
+	}{
+		{"exists first", "bbb", 0},
+		{"exists middle", "eee", 2},
+		{"exists last", "hhh", 5},
+		{"not exists last", "zzz", 6},
+		{"not exists first", "aaa", 0},
+		{"not exists mid", "ddd", 2},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			got := bytesutil.SearchBytes(in, []byte(test.x))
+			if got != test.exp {
+				t.Errorf("got %t, expected %t", got, test.exp)
+			}
+		})
+	}
+}
+
+func TestContains(t *testing.T) {
+	in := toByteSlices("bbb", "ccc", "eee", "fff", "ggg", "hhh")
+	tests := []struct {
+		name string
+		x    string
+		exp  bool
+	}{
+		{"exists first", "bbb", true},
+		{"exists middle", "eee", true},
+		{"exists last", "hhh", true},
+		{"not exists last", "zzz", false},
+		{"not exists first", "aaa", false},
+		{"not exists mid", "ddd", false},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			got := bytesutil.Contains(in, []byte(test.x))
+			if got != test.exp {
+				t.Errorf("got %t, expected %t", got, test.exp)
+			}
+		})
+	}
+}
+
 func TestPack_WidthOne_One(t *testing.T) {
 	a := make([]byte, 8)
 
@@ -128,5 +176,45 @@ func TestPack_WidthOne_LastFill(t *testing.T) {
 		if got, exp := a[i], v; got != exp {
 			t.Fatalf("value mismatch: a[%d] = %v, exp %v", i, got, exp)
 		}
+	}
+}
+
+func BenchmarkContains_True(b *testing.B) {
+	var in [][]byte
+	for i := 'a'; i <= 'z'; i++ {
+		in = append(in, []byte(strings.Repeat(string(i), 3)))
+	}
+	for i := 0; i < b.N; i++ {
+		bytesutil.Contains(in, []byte("xxx"))
+	}
+}
+
+func BenchmarkContains_False(b *testing.B) {
+	var in [][]byte
+	for i := 'a'; i <= 'z'; i++ {
+		in = append(in, []byte(strings.Repeat(string(i), 3)))
+	}
+	for i := 0; i < b.N; i++ {
+		bytesutil.Contains(in, []byte("a"))
+	}
+}
+
+func BenchmarkSearchBytes_Exists(b *testing.B) {
+	var in [][]byte
+	for i := 'a'; i <= 'z'; i++ {
+		in = append(in, []byte(strings.Repeat(string(i), 3)))
+	}
+	for i := 0; i < b.N; i++ {
+		bytesutil.SearchBytes(in, []byte("xxx"))
+	}
+}
+
+func BenchmarkSearchBytes_NotExits(b *testing.B) {
+	var in [][]byte
+	for i := 'a'; i <= 'z'; i++ {
+		in = append(in, []byte(strings.Repeat(string(i), 3)))
+	}
+	for i := 0; i < b.N; i++ {
+		bytesutil.SearchBytes(in, []byte("a"))
 	}
 }