--- a/tsdb/index/inmem/inmem_test.go
+++ b/tsdb/index/inmem/inmem_test.go
@@ -21,9 +21,9 @@ func createData(lo, hi int) (keys, names [][]byte, tags []models.Tags) {
 }
 
 func BenchmarkShardIndex_CreateSeriesListIfNotExists_MaxValuesExceeded(b *testing.B) {
-	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo")}
+	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo", nil)}
 	opt.Config.MaxValuesPerTag = 10
-	si := inmem.NewShardIndex(1, "foo", "bar", opt)
+	si := inmem.NewShardIndex(1, "foo", "bar", nil, opt)
 	si.Open()
 	keys, names, tags := createData(0, 10)
 	si.CreateSeriesListIfNotExists(keys, names, tags)
@@ -37,9 +37,9 @@ func BenchmarkShardIndex_CreateSeriesListIfNotExists_MaxValuesExceeded(b *testin
 }
 
 func BenchmarkShardIndex_CreateSeriesListIfNotExists_MaxValuesNotExceeded(b *testing.B) {
-	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo")}
+	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo", nil)}
 	opt.Config.MaxValuesPerTag = 100000
-	si := inmem.NewShardIndex(1, "foo", "bar", opt)
+	si := inmem.NewShardIndex(1, "foo", "bar", nil, opt)
 	si.Open()
 	keys, names, tags := createData(0, 10)
 	si.CreateSeriesListIfNotExists(keys, names, tags)
@@ -53,8 +53,8 @@ func BenchmarkShardIndex_CreateSeriesListIfNotExists_MaxValuesNotExceeded(b *tes
 }
 
 func BenchmarkShardIndex_CreateSeriesListIfNotExists_NoMaxValues(b *testing.B) {
-	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo")}
-	si := inmem.NewShardIndex(1, "foo", "bar", opt)
+	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo", nil)}
+	si := inmem.NewShardIndex(1, "foo", "bar", nil, opt)
 	si.Open()
 	keys, names, tags := createData(0, 10)
 	si.CreateSeriesListIfNotExists(keys, names, tags)
@@ -68,10 +68,10 @@ func BenchmarkShardIndex_CreateSeriesListIfNotExists_NoMaxValues(b *testing.B) {
 }
 
 func BenchmarkShardIndex_CreateSeriesListIfNotExists_MaxSeriesExceeded(b *testing.B) {
-	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo")}
+	opt := tsdb.EngineOptions{InmemIndex: inmem.NewIndex("foo", nil)}
 	opt.Config.MaxValuesPerTag = 0
 	opt.Config.MaxSeriesPerDatabase = 10
-	si := inmem.NewShardIndex(1, "foo", "bar", opt)
+	si := inmem.NewShardIndex(1, "foo", "bar", nil, opt)
 	si.Open()
 	keys, names, tags := createData(0, 10)
 	si.CreateSeriesListIfNotExists(keys, names, tags)